Fix final comments in PR #2794 - Minor changes, a lot of TODOs.
